feat(media): implement media duration extraction and processing for lectures and lessons
