+++ app/models/lesson.rb
@@ -1,4 +1,6 @@
 class Lesson < ApplicationRecord
+  include MediaHandler
+
   belongs_to :series
 
   validates :title, presence: true, uniqueness: true
@@ -9,9 +11,6 @@ class Lesson < ApplicationRecord
   has_one_attached :optimized_audio, service: Rails.application.config.public_storage
 
   has_rich_text :content
-
-  after_save :extract_media_duration
-  after_commit :process_media_files, on: [ :create, :update ]
 
   # Scopes
   scope :recent, -> { order(published_date: :desc) }
@@ -28,16 +27,8 @@ class Lesson < ApplicationRecord
     [ "series" ]
   end
 
-  def video?
-    video.attached?
-  end
-
   def media_type
       video? ? "video" : "audio"
-  end
-
-  def audio?
-    audio.attached?
   end
 
   def audio_url
@@ -47,30 +38,5 @@ class Lesson < ApplicationRecord
 
   def series_title
     series&.title
-  end
-
-  private
-
-  def process_media_files
-    if video?
-      VideoProcessingJob.perform_later(self)
-    end
-
-    if audio?
-      AudioOptimizationJob.perform_later(self)
-    end
-  end
-
-  def extract_media_duration
-    return unless audio.attached? || video.attached?
-
-    media_file = video.attached? ? video : audio
-
-    media_file.open do |file|
-      movie = FFMPEG::Movie.new(file.path)
-      update_column(:duration, movie.duration.to_i) if movie.duration
-    end
-  rescue => e
-    Rails.logger.error "Failed to extract duration: #{e.message}"
   end
 end
